fix(theme): address theme editor review issues
