+++ src/i18n/locales/en.json
@@ -386,8 +386,18 @@
     "themeColorForeground": "Foreground",
     "themeColorCard": "Card",
     "themeColorPrimary": "Primary",
+    "themeColorPrimaryForeground": "Primary text",
     "themeColorSecondary": "Secondary",
+    "themeColorSecondaryForeground": "Secondary text",
     "themeColorAccent": "Accent",
+    "themeColorAccentForeground": "Accent text",
+    "themeColorCardForeground": "Card text",
+    "themeColorPopover": "Popover",
+    "themeColorPopoverForeground": "Popover text",
+    "themeColorMuted": "Muted",
+    "themeColorMutedForeground": "Muted text",
+    "themeColorDestructive": "Destructive",
+    "themeColorDestructiveForeground": "Destructive text",
     "themeColorBorder": "Border",
     "themeColorInput": "Input",
     "themeColorRing": "Focus ring",
