+++ src/i18n/locales/ja.json
@@ -386,8 +386,18 @@
     "themeColorForeground": "前景色",
     "themeColorCard": "カード",
     "themeColorPrimary": "Primary",
+    "themeColorPrimaryForeground": "Primary text",
     "themeColorSecondary": "Secondary",
+    "themeColorSecondaryForeground": "Secondary text",
     "themeColorAccent": "Accent",
+    "themeColorAccentForeground": "Accent text",
+    "themeColorCardForeground": "Card text",
+    "themeColorPopover": "Popover",
+    "themeColorPopoverForeground": "Popover text",
+    "themeColorMuted": "Muted",
+    "themeColorMutedForeground": "Muted text",
+    "themeColorDestructive": "Destructive",
+    "themeColorDestructiveForeground": "Destructive text",
     "themeColorBorder": "Border",
     "themeColorInput": "Input",
     "themeColorRing": "フォーカスリング",
