+++ src/i18n/locales/zh.json
@@ -386,8 +386,18 @@
     "themeColorForeground": "前景色",
     "themeColorCard": "卡片色",
     "themeColorPrimary": "主色",
+    "themeColorPrimaryForeground": "主色文字",
     "themeColorSecondary": "次级色",
+    "themeColorSecondaryForeground": "次级色文字",
     "themeColorAccent": "强调色",
+    "themeColorAccentForeground": "强调色文字",
+    "themeColorCardForeground": "卡片文字",
+    "themeColorPopover": "弹出层色",
+    "themeColorPopoverForeground": "弹出层文字",
+    "themeColorMuted": "弱化色",
+    "themeColorMutedForeground": "弱化色文字",
+    "themeColorDestructive": "危险色",
+    "themeColorDestructiveForeground": "危险色文字",
     "themeColorBorder": "边框色",
     "themeColorInput": "输入框",
     "themeColorRing": "焦点环",
